requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
motor>=3.7.1
pymongo>=4.10.1
certifi>=2025.11.12
pyopenssl>=25.3.0

# Vector Search & Embeddings
fastembed==0.3.6

# LLM Integration
openai==1.59.5

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
python-magic-bin==0.4.14

# Utilities
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
